'Protecting Title' becomes tri-state
The file property 'protectig titles' (applies to book titles
and series as well) becomes tri-state when loading and saving
.bib files:
- Remove all protective curly brackets, i.e. "{Abc}" becomes "Abc"
- Add protectie curly brackets, i.e. "Abc" becomes "{Abc}"
- Do not touch any protective titles

When loading .bib files, it is counted how many titles are
protected and how many are not.
If the vast majority is protected, the whole file will be flagged
as 'titles are to be protected' which gets applied to *all* titles
when saving the file.
If the vast majority of titles are not protected, the flag is set
to the opposity and for the minority of protected titles, the
protection gets removed when saving.
If some titles are protected and some are not, the title protection
is set to 'undecided'.

+++ src/data/file.cpp
@@ -81,7 +81,7 @@ public:
         properties.insert(File::QuoteComment, (Preferences::QuoteComment)configGroup.readEntry(Preferences::keyQuoteComment, (int)Preferences::defaultQuoteComment));
         properties.insert(File::KeywordCasing, (KBibTeX::Casing)configGroup.readEntry(Preferences::keyKeywordCasing, (int)Preferences::defaultKeywordCasing));
         properties.insert(File::NameFormatting, configGroup.readEntry(Preferences::keyPersonNameFormatting, QString()));
-        properties.insert(File::ProtectCasing, configGroup.readEntry(Preferences::keyProtectCasing, Preferences::defaultProtectCasing));
+        properties.insert(File::ProtectCasing, configGroup.readEntry(Preferences::keyProtectCasing, (int)Preferences::defaultProtectCasing));
         properties.insert(File::ListSeparator, configGroup.readEntry(Preferences::keyListSeparator, Preferences::defaultListSeparator));
     }
